+++ cmd/api/main.go
@@ -7,10 +7,14 @@ import (
 	"time"
 
 	defaultresolver "github.com/Chatterino/api/internal/resolvers/default"
+	"github.com/Chatterino/api/internal/routes/twitchemotes"
+	"github.com/Chatterino/api/internal/twitchapiclient"
+	"github.com/Chatterino/api/pkg/cache"
 	"github.com/Chatterino/api/pkg/config"
 	"github.com/Chatterino/api/pkg/resolver"
 	"github.com/Chatterino/api/pkg/thumbnail"
 	"github.com/go-chi/chi/v5"
+	"github.com/go-chi/chi/v5/middleware"
 )
 
 var (
@@ -19,12 +23,10 @@ var (
 	}
 	startTime = time.Now()
 
-	cfg = config.New()
-
 	prefix string
 )
 
-func mountRouter(r *chi.Mux) *chi.Mux {
+func mountRouter(r *chi.Mux, cfg config.APIConfig) *chi.Mux {
 	if cfg.BaseURL == "" {
 		log.Printf("Listening on %s (Prefix=%s, BaseURL=%s)\n", cfg.BindAddress, prefix, cfg.BaseURL)
 		return r
@@ -60,14 +62,26 @@ func listen(bind string, router *chi.Mux) {
 }
 
 func main() {
+	cfg := config.New()
+
 	resolver.InitializeStaticResponses(cfg)
 	thumbnail.InitializeConfig(cfg)
 
 	router := chi.NewRouter()
 
-	handleTwitchEmotes(router)
-	handleHealth(router)
-	defaultresolver.Initialize(router, cfg)
+	// Strip trailing slashes from API requests
+	router.Use(middleware.StripSlashes)
 
-	listen(cfg.BindAddress, mountRouter(router))
+	var helixUsernameCache *cache.Cache
+
+	helixClient, helixUsernameCache, err := twitchapiclient.New(cfg)
+	if err != nil {
+		log.Printf("[Twitch] %s\n", err.Error())
+	}
+
+	twitchemotes.Initialize(router, helixClient, helixUsernameCache)
+	handleHealth(router)
+	defaultresolver.Initialize(router, cfg, helixClient)
+
+	listen(cfg.BindAddress, mountRouter(router, cfg))
 }