+++ internal/resolvers/seventv/run.go
@@ -1,0 +1,19 @@
+package seventv
+
+import (
+	"encoding/json"
+	"net/http"
+	"net/url"
+)
+
+func run(url *url.URL, r *http.Request) ([]byte, error) {
+	matches := emotePathRegex.FindStringSubmatch(url.Path)
+	if len(matches) != 2 {
+		return nil, errInvalidSevenTVEmotePath
+	}
+
+	emoteHash := matches[1]
+
+	apiResponse := emoteCache.Get(emoteHash, r)
+	return json.Marshal(apiResponse)
+}