+++ internal/twitchapiclient/token_refresh.go
@@ -1,4 +1,4 @@
-package twitch
+package twitchapiclient
 
 import (
 	"log"