+++ pkg/utils/bit.go
@@ -1,0 +1,6 @@
+package utils
+
+// HasBits checks if sum contains bit by performing a bitwise AND operation between values
+func HasBits(sum int32, bit int32) bool {
+	return (sum & bit) == bit
+}